♻️ refactor: post page 서버사이드로 변경경
