+++ src/utils/api/supabase_api/getUser.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from '@/utils/supabase/server';
 
 export const getUser = async () => {
   const supabase = await createClient();
